requirements.txt
arrow==0.15.6
binaryornot==0.4.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
future==0.18.2
idna==2.9
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
passlib==1.7.2
poyo==0.5.0
python-dateutil==2.8.1
python-slugify==4.0.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
text-unidecode==1.3
urllib3==1.25.9
whichcraft==0.6.1